[networks] Fix import loop
